+++ app/screens/settings/advanced/index.tsx
@@ -6,13 +6,14 @@ import {useIntl} from 'react-intl';
 import {View, TouchableOpacity} from 'react-native';
 import {Edge, SafeAreaView} from 'react-native-safe-area-context';
 
-import OptionItem from '@components/option_item';
 import {useTheme} from '@context/theme';
 import useAndroidHardwareBackHandler from '@hooks/android_back_handler';
 import {popTopScreen} from '@screens/navigation';
 import {deleteFileCache, getAllFilesInCachesDirectory, getFormattedFileSize} from '@utils/file';
 import {preventDoubleTap} from '@utils/tap';
 import {changeOpacity, makeStyleSheetFromTheme} from '@utils/theme';
+
+import SettingOption from '../setting_option';
 
 import type {ReadDirItem} from 'react-native-fs';
 
@@ -94,7 +95,7 @@ const AdvancedSettings = ({componentId}: AdvancedSettingsProps) => {
                     disabled={disabled}
                     activeOpacity={disabled ? 0 : 1}
                 >
-                    <OptionItem
+                    <SettingOption
                         containerStyle={styles.itemStyle}
                         destructive={true}
                         label={intl.formatMessage({id: 'advanced_settings.delete_data', defaultMessage: 'Delete Documents & Data'})}
